+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 * Change order in structure view to match source file and sectioning level 
 * Add command redefinitions to command definition filter in structure view
+* Add support for automatic language injection on the minted environment
+* Add support for automatic detection of custom command aliases for include commands
 
 ### Fixed
 
@@ -12,7 +14,6 @@
 
 ### Added
 
-* Add support for automatic language injection on the minted environment
 * Add support for DeclareGraphicsExtensions
 * Add inspection to warn about a missing reference for a glossary occurrence
 * Do not fold sections in a command definition